requirements.txt
APScheduler==3.6.3
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
pycodestyle==2.6.0
pygame==2.0.0
Pygments==2.8.1
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2021.1
pywin32==300
pyzmq==22.0.3
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
wcwidth==0.2.5
win10toast==0.9
